Major refactoring of libvipster
- removed AtomProper
- reworked AtomPropRef
  now takes Memberpointer as template-argument,
  better SFINAE, pointers to value stay in Atom
  -> more/direct control
- introduced bitset/PseEntry pointers to Atom
  direct access to PseEntry, joined fix and hidden
  with space for more properties
- reworked Step
  collected Bond and Cell data in structs
  Split common functionality to StepBase
  Implemented as CRTP-Base in Step
  Paved way for StepSel with own Atom-interface
  while reducing amount of code

INCOMPLETE:
- check Bond-setting
- removed/broke tests, reimplement them

+++ libvipster/ioplugins/xyz.cpp
@@ -14,7 +14,7 @@ std::shared_ptr<IO::BaseData> XYZParser(std::string name, std::ifstream &file)
     std::string line;
     while (std::getline(file, line)) {
         std::stringstream natline{line};
-        int nat;
+        size_t nat;
         natline >> nat;
         if (natline.fail()) {
             if (!m.getNstep()) throw IOError("XYZ: Failed to parse nat");
